move files to packages, remove unused import
