+++ src/main/kotlin/org/wycliffeassociates/sourceaudio/upload/AppModule.kt
@@ -1,3 +1,5 @@
+package org.wycliffeassociates.sourceaudio.upload
+
 import io.ktor.application.*
 import io.ktor.features.CORS
 import io.ktor.features.CallLogging
@@ -11,15 +13,9 @@ import io.ktor.request.acceptLanguage
 import io.ktor.request.receiveMultipart
 import org.thymeleaf.templateresolver.ClassLoaderTemplateResolver
 import org.wycliffeassociates.sourceaudio.routing.SupportedLanguages
-import org.wycliffeassociates.sourceaudio.upload.RequestHandler
-import org.wycliffeassociates.sourceaudio.upload.TemplateModel
-import org.wycliffeassociates.sourceaudio.upload.model.CompressedExtensions
-import org.wycliffeassociates.sourceaudio.upload.model.Groupings
-import org.wycliffeassociates.sourceaudio.upload.model.MediaQuality
-import org.wycliffeassociates.sourceaudio.upload.model.UncompressedExtensions
-import java.util.*
+import java.util.Locale
 
-fun Application.module() {
+fun Application.appModule() {
     install(DefaultHeaders)
     install(CORS) {
         anyHost()
@@ -36,8 +32,8 @@ fun Application.module() {
     install(Routing) {
         routing {
             static("static") {
-                files("src/css")
-                files("src/js")
+                resources("css")
+                resources("js")
             }
         }
         get("/") {
@@ -65,7 +61,7 @@ fun Application.module() {
     }
 }
 
-fun getPreferredLocale(languageRanges: List<Locale.LanguageRange>): Locale {
+private fun getPreferredLocale(languageRanges: List<Locale.LanguageRange>): Locale {
     var language = ""
     for (range in languageRanges) {
         language = range.range.split("-")[0]
@@ -75,3 +71,5 @@ fun getPreferredLocale(languageRanges: List<Locale.LanguageRange>): Locale {
 
     return Locale(language)
 }
+
+val Application.envKind get() = environment.config.property("ktor.environment").getString()
