+++ src/main/kotlin/org/wycliffeassociates/sourceaudio/upload/Startup.kt
@@ -1,0 +1,16 @@
+package org.wycliffeassociates.sourceaudio.upload
+
+import io.ktor.application.Application
+import io.ktor.server.engine.embeddedServer
+import io.ktor.server.netty.Netty
+
+const val port: Int = 4567
+
+fun main(args: Array<String>) {
+    embeddedServer(
+        Netty,
+        port = port,
+        watchPaths = listOf("KotlinWeb"),
+        module = Application::appModule
+    ).start(true)
+}
